+++ aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/strategy/sampling/CentralizedSamplingStrategy.java
@@ -91,6 +91,9 @@ public class CentralizedSamplingStrategy implements SamplingStrategy {
     }
 
     @Override
+    /**
+     * Shutdown all poller threads immediately regardless of the pending work for clean exit.
+     */
     public void shutdown() {
         rulePoller.shutdown();
         targetPoller.shutdown();