+++ aws-xray-recorder-sdk-core/src/test/java/com/amazonaws/xray/config/DaemonConfigurationTest.java
@@ -1,6 +1,10 @@
 package com.amazonaws.xray.config;
 
-import org.junit.*;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.FixMethodOrder;
+import org.junit.Rule;
+import org.junit.Test;
 import org.junit.contrib.java.lang.system.EnvironmentVariables;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.runners.MethodSorters;